# Fixes #1000 - Speed up builds