--- a/package.json
+++ b/package.json
@@ -39,6 +39,7 @@
     "karma-verbose-reporter": "0.0.6",
     "karma-webpack": "^3.0.0",
     "mocha": "5.2.0",
+    "parallel-webpack": "2.3.0",
     "prettier": "1.15.3",
     "sinon": "1.15.4",
     "terser-webpack-plugin": "^1.2.1",