# Step 1. Avoid accessing stats (doesn't help much)

According to https://webpack.js.org/guides/build-performance/ accessing fields on the stats object can be expensive. So, I changed things here to only access it if `stats.hasErrors()` is true, or if you pass in an environment variable:

```sh
env STATS=1 npm run build
```

By default, stats won't be shown. Instead, you'll see this in the build output:

```
[15:30:44] Starting 'build:webpack.dev'...

Not showing webpack stats; run with `env STATS=1` to show

[15:31:07] Finished 'build:webpack.dev' after 23 s

Not showing webpack stats; run with `env STATS=1` to show

[15:31:11] Finished 'build:webpack.prod' after 27 s
```

In testing this, I found a bug in the original code which explains why the stats output is not in color. I was calling `stats.stats.toString({colors: true})` but I needed to call `stats.toString({colors: true})`. It still "worked", but not in color.

The bottom-line? Sadly this doesn't trim anything much noticeable off the build time, but at least it fixes a bug and makes the output cleaner, so going with it.

# Step 2. Use a whitelist instead of a blacklist (doesn't help much)

# Step 3. Use parallel-webpack (helps a lot)

So we were already doing some amount of parallelization here (Gulp tasks in parallel, minification in parallel), but because a lot of this is I/O-bound we still have idle CPU resources sitting around which could be compiling our bundles for us.

parallel-webpack uses child worker processes, and we can actually saturate the CPU with it, at least on my laptop, dropping build times in half (from about 29s to 14s).

Here's evidence that we actually make use of the CPU now; previously we only used about 10%:

![cpu](https://user-images.githubusercontent.com/7074/51846859-b982f500-231a-11e9-8690-297fbf68552e.png)

Build output looks like this:

```
[16:22:21] Starting 'build:webpack.prod'...
[16:22:21] Starting 'build:webpack.dev'...

Not showing webpack stats; run with `env STATS=1` to show

[WEBPACK] Building 4 targets
[WEBPACK] Building 4 targets
[WEBPACK] Started building alloy-editor-no-ckeditor.js
[WEBPACK] Started building alloy-editor-core.js
[WEBPACK] Started building alloy-editor-all.js
[WEBPACK] Started building alloy-editor-no-react.js
[WEBPACK] Started building alloy-editor-all-min.js
[WEBPACK] Started building alloy-editor-no-ckeditor-min.js
[WEBPACK] Started building alloy-editor-core-min.js
[WEBPACK] Started building alloy-editor-no-react-min.js
[WEBPACK] Finished building alloy-editor-core.js within 9.037 seconds
[WEBPACK] Finished building alloy-editor-no-ckeditor.js within 9.588 seconds
[WEBPACK] Finished building alloy-editor-no-react.js within 9.632 seconds
[WEBPACK] Finished building alloy-editor-core-min.js within 9.475 seconds
[WEBPACK] Finished building alloy-editor-all.js within 9.984 seconds
[WEBPACK] Finished build after 10.836 seconds
[16:22:32] Finished 'build:webpack.dev' after 11 s
[WEBPACK] Finished building alloy-editor-no-ckeditor-min.js within 9.974 seconds
[WEBPACK] Finished building alloy-editor-no-react-min.js within 10.786 seconds
[WEBPACK] Finished building alloy-editor-all-min.js within 11.275 seconds
[WEBPACK] Finished build after 12.389 seconds
[16:22:34] Finished 'build:webpack.prod' after 13 s
[16:22:34] Finished 'build' after 13 s
[16:22:34] Finished 'default' after 13 s
```

As you can see, I kept the `env STATS=1` pattern from Step 1.

# Testing

- `npm run test` still passes
- `npm run start` and test in browser
- compare "dist" results

## "dist" before

```
   882296 28 Jan 16:49 dist/alloy-editor/alloy-editor-all-min.js
  5517088 28 Jan 16:49 dist/alloy-editor/alloy-editor-all.js
   278868 28 Jan 16:49 dist/alloy-editor/alloy-editor-core-min.js
  2069445 28 Jan 16:49 dist/alloy-editor/alloy-editor-core.js
   390120 28 Jan 16:49 dist/alloy-editor/alloy-editor-no-ckeditor-min.js
  4213528 28 Jan 16:49 dist/alloy-editor/alloy-editor-no-ckeditor.js
   771364 28 Jan 16:49 dist/alloy-editor/alloy-editor-no-react-min.js
  3376525 28 Jan 16:49 dist/alloy-editor/alloy-editor-no-react.js
```

## "dist" after

```
   882102 28 Jan 16:48 dist/alloy-editor/alloy-editor-all-min.js
  5517345 28 Jan 16:48 dist/alloy-editor/alloy-editor-all.js
   278868 28 Jan 16:48 dist/alloy-editor/alloy-editor-core-min.js
  2069445 28 Jan 16:48 dist/alloy-editor/alloy-editor-core.js
   390120 28 Jan 16:48 dist/alloy-editor/alloy-editor-no-ckeditor-min.js
  4213528 28 Jan 16:48 dist/alloy-editor/alloy-editor-no-ckeditor.js
   771170 28 Jan 16:48 dist/alloy-editor/alloy-editor-no-react-min.js
  3376786 28 Jan 16:48 dist/alloy-editor/alloy-editor-no-react.js
```

Closes: #1000